+++ include/magicxx/magic.hpp
@@ -4,11 +4,11 @@
 #ifndef MAGIC_HPP
 #define MAGIC_HPP
 
-#include <algorithm>
 #include <bitset>
 #include <expected>
 #include <map>
 #include <memory>
+#include <optional>
 #include <string_view>
 #include <vector>
 
@@ -22,9 +22,10 @@ namespace recognition {
  *
  * @brief The magic class provides a C++ wrapper over the Magic Number Recognition Library.
  *
- * @note  The magic class is used to identify the type of a file, if the following steps have been completed;
- *        1. magic must be opened.
- *        2. A magic database file must be loaded.
+ * @note  To use the magic class for file type identification, ensure the following:
+ *        1. The magic instance must be opened.
+ *        2. A magic database file must be successfully loaded.
+ *        Only after these steps is the instance considered valid for identifying file types.
  */
 class magic {
 public:
@@ -137,6 +138,9 @@ public:
 
     /**
      * @brief Construct magic without opening it.
+     * 
+     * @note This magic is not valid for identifying file types
+     *       until it is opened and a magic database file is loaded.
      */
     magic() noexcept;
 
@@ -158,6 +162,21 @@ public:
     );
 
     /**
+     * @brief Construct magic, open it using the flags and load the magic database file, noexcept version.
+     *
+     * @param[in] flags_mask        One of the flags enums or bitwise or of the flags enums.
+     * @param[in] tag               Tag for non-throwing overload.
+     * @param[in] database_file     The path of magic database file, default is default_database_file.
+     *
+     * @note load_database_file() adds “.mgc” to the database filename as appropriate.
+     */
+    magic(
+        flags_mask_t                           flags_mask,
+        [[maybe_unused]] const std::nothrow_t& tag,
+        const std::filesystem::path& database_file = default_database_file
+    ) noexcept;
+
+    /**
      * @brief Construct magic, open it using the flags and load the magic database file.
      *
      * @param[in] flags_container   Flags.
@@ -175,9 +194,26 @@ public:
     );
 
     /**
+     * @brief Construct magic, open it using the flags and load the magic database file, noexcept version.
+     *
+     * @param[in] flags_container   Flags.
+     * @param[in] tag               Tag for non-throwing overload.
+     * @param[in] database_file     The path of magic database file, default is default_database_file.
+     *
+     * @note load_database_file() adds “.mgc” to the database filename as appropriate.
+     */
+    magic(
+        const flags_container_t&               flags_container,
+        [[maybe_unused]] const std::nothrow_t& tag,
+        const std::filesystem::path& database_file = default_database_file
+    ) noexcept;
+
+    /**
      * @brief Move construct magic.
      *
-     * @note other is valid as a default constructed magic.
+     * @note After move construction, the moved-from object (other) is closed.
+     *       It cannot be used for identifying file types until 
+     *       it is reopened and a magic database file is loaded.
      */
     magic(magic&& other) noexcept;
 
@@ -189,7 +225,9 @@ public:
     /**
      * @brief Move assign to this magic.
      *
-     * @note other is valid as a default constructed magic.
+     * @note After move construction, the moved-from object (other) is closed.
+     *       It cannot be used for identifying file types until
+     *       it is reopened and a magic database file is loaded.
      */
     magic& operator=(magic&& other) noexcept;
 
@@ -204,9 +242,12 @@ public:
     ~magic();
 
     /**
-     * @brief Used for testing whether magic is open or closed.
+     * @brief Used for testing whether magic is valid
+     *        for identifying file types or not.
      *
-     * @returns True if magic is open, false otherwise.
+     * @returns True if the magic is valid, i.e.,
+     *          it is open and a magic database file is loaded;
+     *          false otherwise.
      */
     [[nodiscard]] operator bool() const noexcept;
 
@@ -225,7 +266,8 @@ public:
     /**
      * @brief Close magic.
      *
-     * @note magic is valid as a default constructed magic.
+     * @note magic cannot be used for identifying file types until
+     *       it is reopened and a magic database file is loaded.
      */
     void close() noexcept;
 
@@ -254,6 +296,17 @@ public:
     [[nodiscard]] flags_container_t get_flags() const;
 
     /**
+     * @brief Get the flags of magic, noexcept version.
+     *
+     * @param[in] tag               Tag for non-throwing overload.
+     *
+     * @returns flags_container_t or std::nullopt if magic is closed.
+     */
+    [[nodiscard]] std::optional<flags_container_t> get_flags(
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) const noexcept;
+
+    /**
      * @brief Get the value of a parameter of magic.
      *
      * @param[in] parameter         One of the parameters enum.
@@ -265,13 +318,37 @@ public:
     [[nodiscard]] std::size_t get_parameter(parameters parameter) const;
 
     /**
+     * @brief Get the value of a parameter of magic, noexcept version.
+     *
+     * @param[in] parameter         One of the parameters enum.
+     * @param[in] tag               Tag for non-throwing overload.
+     *
+     * @returns Value of the desired parameter or std::nullopt if magic is closed.
+     */
+    [[nodiscard]] std::optional<std::size_t> get_parameter(
+        parameters                             parameter,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) const noexcept;
+
+    /**
      * @brief Get the values ​​of all parameters of magic.
      *
-     * @returns <parameters, value> map.
+     * @returns <parameter, value> map.
      *
      * @throws magic_is_closed      if magic is closed.
      */
     [[nodiscard]] parameter_value_map_t get_parameters() const;
+
+    /**
+     * @brief Get the values ​​of all parameters of magic, noexcept version.
+     *
+     * @param[in] tag               Tag for non-throwing overload.
+     *
+     * @returns <parameter, value> map or an empty map if magic is closed.
+     */
+    [[nodiscard]] parameter_value_map_t get_parameters(
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) const noexcept;
 
     /**
      * @brief Get the version of the Magic Number Recognition Library.
@@ -299,12 +376,13 @@ public:
      * @brief Identify the type of a file, noexcept version.
      *
      * @param[in] path              The path of the file.
+     * @param[in] tag               Tag for non-throwing overload.
      *
      * @returns The type of the file or the error message.
      */
     [[nodiscard]] expected_file_type_t identify_file(
-        const std::filesystem::path& path,
-        std::nothrow_t
+        const std::filesystem::path&           path,
+        [[maybe_unused]] const std::nothrow_t& tag
     ) const noexcept;
 
     /**
@@ -324,38 +402,23 @@ public:
         const std::filesystem::path&       directory,
         std::filesystem::directory_options option = std::filesystem::
             directory_options::follow_directory_symlink
-    ) const
-    {
-        return identify_files_impl(
-            std::filesystem::recursive_directory_iterator{directory, option}
-        );
-    }
+    ) const;
 
     /**
      * @brief Identify the types of all files in a directory, noexcept version.
      *
      * @param[in] directory         The path of the directory.
+     * @param[in] tag               Tag for non-throwing overload.
      * @param[in] option            The directory iteration option, default is follow_directory_symlink.
      *
      * @returns The types of each file as a map.
      */
     [[nodiscard]] expected_types_of_files_t identify_files(
-        const std::filesystem::path& directory,
-        std::nothrow_t,
-        std::filesystem::directory_options option = std::filesystem::
+        const std::filesystem::path&           directory,
+        [[maybe_unused]] const std::nothrow_t& tag,
+        std::filesystem::directory_options     option = std::filesystem::
             directory_options::follow_directory_symlink
-    ) const noexcept
-    {
-        std::error_code error_code{};
-        return identify_files_impl(
-            std::filesystem::recursive_directory_iterator{
-                directory,
-                option,
-                error_code
-            },
-            std::nothrow
-        );
-    }
+    ) const noexcept;
 
     /**
      * @brief Identify the types of files.
@@ -371,25 +434,20 @@ public:
      */
     [[nodiscard]] types_of_files_t identify_files(
         const file_concepts::file_container auto& files
-    ) const
-    {
-        return identify_files_impl(files);
-    }
+    ) const;
 
     /**
      * @brief Identify the types of files, noexcept version.
      *
      * @param[in] files             The container that holds the paths of the files.
+     * @param[in] tag               Tag for non-throwing overload.
      *
      * @returns The types of each file as a map.
      */
     [[nodiscard]] expected_types_of_files_t identify_files(
         const file_concepts::file_container auto& files,
-        std::nothrow_t
-    ) const noexcept
-    {
-        return identify_files_impl(files, std::nothrow);
-    }
+        [[maybe_unused]] const std::nothrow_t&    tag
+    ) const noexcept;
 
     /**
      * @brief Used for testing whether magic is open or closed.
@@ -397,6 +455,16 @@ public:
      * @returns True if magic is open, false otherwise.
      */
     [[nodiscard]] bool is_open() const noexcept;
+
+    /**
+     * @brief Used for testing whether magic is valid
+     *        for identifying file types or not.
+     *
+     * @returns True if the magic is valid, i.e.,
+     *          it is open and a magic database file is loaded;
+     *          false otherwise.
+     */
+    [[nodiscard]] bool is_valid() const noexcept;
 
     /**
      * @brief Load a magic database file.
@@ -415,6 +483,21 @@ public:
     );
 
     /**
+     * @brief Load a magic database file, noexcept version.
+     *
+     * @param[in] tag               Tag for non-throwing overload.
+     * @param[in] database_file     The path of the magic database file, default is default_database_file.
+     *
+     * @returns True on success, false otherwise.
+     *
+     * @note load_database_file() adds “.mgc” to the database filename as appropriate.
+     */
+    bool load_database_file(
+        [[maybe_unused]] const std::nothrow_t& tag,
+        const std::filesystem::path& database_file = default_database_file
+    ) noexcept;
+
+    /**
      * @brief Open magic using the flags.
      *
      * @param[in] flags_mask        One of the flags enums or bitwise or of the flags enums.
@@ -422,8 +505,25 @@ public:
      * @throws magic_open_error     if opening magic fails.
      *
      * @note If magic is open, it will be reopened using the flags after closing it.
+     * @note A magic database file must be loaded after opening magic.
      */
     void open(flags_mask_t flags_mask);
+
+    /**
+     * @brief Open magic using the flags, noexcept version.
+     *
+     * @param[in] flags_mask        One of the flags enums or bitwise or of the flags enums.
+     * @param[in] tag               Tag for non-throwing overload.
+     *
+     * @returns True on success, false otherwise.
+     *
+     * @note If magic is open, it will be reopened using the flags after closing it.
+     * @note A magic database file must be loaded after opening magic.
+     */
+    bool open(
+        flags_mask_t                           flags_mask,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) noexcept;
 
     /**
      * @brief Open magic using the flags.
@@ -433,8 +533,25 @@ public:
      * @throws magic_open_error     if opening magic fails.
      *
      * @note If magic is open, it will be reopened using the flags after closing it.
+     * @note A magic database file must be loaded after opening magic.
      */
     void open(const flags_container_t& flags_container);
+
+    /**
+     * @brief Open magic using the flags, noexcept version.
+     *
+     * @param[in] flags_container   Flags.
+     * @param[in] tag               Tag for non-throwing overload.
+     *
+     * @returns True on success, false otherwise.
+     *
+     * @note If magic is open, it will be reopened using the flags after closing it.
+     * @note A magic database file must be loaded after opening magic.
+     */
+    bool open(
+        const flags_container_t&               flags_container,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) noexcept;
 
     /**
      * @brief Set the flags of magic.
@@ -447,6 +564,19 @@ public:
     void set_flags(flags_mask_t flags_mask);
 
     /**
+     * @brief Set the flags of magicü noexcept version.
+     *
+     * @param[in] flags_mask          One of the flags enums or bitwise or of the flags enums.
+     * @param[in] tag                 Tag for non-throwing overload.
+     *
+     * @returns True on success, false otherwise.
+     */
+    bool set_flags(
+        flags_mask_t                           flags_mask,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) noexcept;
+
+    /**
      * @brief Set the flags of magic.
      *
      * @param[in] flags_container     Flags.
@@ -455,6 +585,19 @@ public:
      * @throws magic_set_flags_error  if setting the flags of magic fails.
      */
     void set_flags(const flags_container_t& flags_container);
+
+    /**
+     * @brief Set the flags of magic, noexcept version.
+     *
+     * @param[in] flags_container     Flags.
+     * @param[in] tag                 Tag for non-throwing overload.
+     *
+     * @returns True on success, false otherwise.
+     */
+    bool set_flags(
+        const flags_container_t&               flags_container,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) noexcept;
 
     /**
      * @brief Set the value of a parameter of magic.
@@ -468,6 +611,21 @@ public:
     void set_parameter(parameters parameter, std::size_t value);
 
     /**
+     * @brief Set the value of a parameter of magic, noexcept version.
+     *
+     * @param[in] parameter           One of the parameters enum.
+     * @param[in] value               The value of the parameter.
+     * @param[in] tag                 Tag for non-throwing overload.
+     *
+     * @returns True on success, false otherwise.
+     */
+    bool set_parameter(
+        parameters                             parameter,
+        std::size_t                            value,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) noexcept;
+
+    /**
      * @brief Set the values of the parameters of magic.
      *
      * @param[in] parameters          Parameters with corresponding values.
@@ -477,32 +635,22 @@ public:
      */
     void set_parameters(const parameter_value_map_t& parameters);
 
+    /**
+     * @brief Set the values of the parameters of magic, noexcept version.
+     *
+     * @param[in] parameters          Parameters with corresponding values.
+     * @param[in] tag                 Tag for non-throwing overload.
+     *
+     * @returns True on success, false otherwise.
+     */
+    bool set_parameters(
+        const parameter_value_map_t&           parameters,
+        [[maybe_unused]] const std::nothrow_t& tag
+    ) noexcept;
+
 private:
     class magic_private;
     std::unique_ptr<magic_private> m_impl;
-
-    [[nodiscard]] types_of_files_t identify_files_impl(
-        const std::ranges::range auto& files
-    ) const
-    {
-        types_of_files_t types_of_files;
-        std::ranges::for_each(files, [&](const std::filesystem::path& file) {
-            types_of_files[file] = identify_file(file);
-        });
-        return types_of_files;
-    }
-
-    [[nodiscard]] expected_types_of_files_t identify_files_impl(
-        const std::ranges::range auto& files,
-        std::nothrow_t
-    ) const noexcept
-    {
-        expected_types_of_files_t expected_types_of_files;
-        std::ranges::for_each(files, [&](const std::filesystem::path& file) {
-            expected_types_of_files[file] = identify_file(file, std::nothrow);
-        });
-        return expected_types_of_files;
-    }
 
     friend std::string to_string(flags);
     friend std::string to_string(parameters);