+++ tests/magic_load_database_file_test.cpp
@@ -11,24 +11,25 @@ using namespace recognition;
 
 struct magic_load_database_file_test : public ::testing::Test {
 protected:
-    std::filesystem::path test_dir = std::filesystem::temp_directory_path()
-                                   / "magic_load_database_file_test";
-    std::filesystem::path invalid_database = test_dir / "invalid_database";
-    magic                 opened_magic{};
+    std::filesystem::path m_test_dir = std::filesystem::temp_directory_path()
+                                     / "magicxx_test";
+    std::filesystem::path m_invalid_database = m_test_dir / "invalid_database";
+    std::filesystem::path m_empty_path;
+    magic                 m_opened_magic{};
 
     void SetUp() override
     {
-        std::filesystem::create_directory(test_dir);
-        std::ofstream file{invalid_database, std::ios::trunc};
+        std::filesystem::create_directory(m_test_dir);
+        std::ofstream file{m_invalid_database, std::ios::trunc};
         file << "test";
         file.flush();
-        opened_magic.open(magic::flags::mime);
+        m_opened_magic.open(magic::flags::mime);
     }
 
     void TearDown() override
     {
-        std::filesystem::remove(invalid_database);
-        std::filesystem::remove(test_dir);
+        std::filesystem::remove(m_invalid_database);
+        std::filesystem::remove(m_test_dir);
     }
 };
 
@@ -39,23 +40,25 @@ TEST_F(magic_load_database_file_test, closed_magic_load_default_database)
 
 TEST_F(magic_load_database_file_test, opened_magic_load_empty_path)
 {
-    EXPECT_THROW(opened_magic.load_database_file({}), empty_path);
+    EXPECT_THROW(m_opened_magic.load_database_file(m_empty_path), empty_path);
 }
 
 TEST_F(magic_load_database_file_test, opened_magic_load_directory)
 {
-    EXPECT_THROW(opened_magic.load_database_file(test_dir), invalid_path);
+    EXPECT_THROW(m_opened_magic.load_database_file(m_test_dir), invalid_path);
 }
 
 TEST_F(magic_load_database_file_test, opened_magic_load_invalid_database)
 {
     EXPECT_THROW(
-        opened_magic.load_database_file(invalid_database),
+        m_opened_magic.load_database_file(m_invalid_database),
         magic_load_error
     );
 }
 
 TEST_F(magic_load_database_file_test, opened_magic_load_default_database)
 {
-    opened_magic.load_database_file(DEFAULT_DATABASE_FILE);
+    EXPECT_FALSE(m_opened_magic.is_valid());
+    m_opened_magic.load_database_file(DEFAULT_DATABASE_FILE);
+    EXPECT_TRUE(m_opened_magic.is_valid());
 }