+++ tests/magic_open_close_test.cpp
@@ -12,6 +12,7 @@ TEST(magic_open_close_test, closed_magic_is_open)
     magic m;
     EXPECT_FALSE(m);
     EXPECT_FALSE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }
 
 TEST(magic_open_close_test, closed_magic_close)
@@ -20,14 +21,16 @@ TEST(magic_open_close_test, closed_magic_close)
     m.close();
     EXPECT_FALSE(m);
     EXPECT_FALSE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }
 
 TEST(magic_open_close_test, opened_magic_is_open)
 {
     magic m;
     m.open(magic::flags::mime);
-    EXPECT_TRUE(m);
+    EXPECT_FALSE(m);
     EXPECT_TRUE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }
 
 TEST(magic_open_close_test, opened_magic_close)
@@ -37,6 +40,7 @@ TEST(magic_open_close_test, opened_magic_close)
     m.close();
     EXPECT_FALSE(m);
     EXPECT_FALSE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }
 
 TEST(magic_open_close_test, opened_magic_open_twice)
@@ -44,8 +48,9 @@ TEST(magic_open_close_test, opened_magic_open_twice)
     magic m;
     m.open(magic::flags::mime);
     m.open(magic::flags::mime_type);
-    EXPECT_TRUE(m);
+    EXPECT_FALSE(m);
     EXPECT_TRUE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }
 
 TEST(magic_open_close_test, opened_magic_close_twice)
@@ -56,6 +61,7 @@ TEST(magic_open_close_test, opened_magic_close_twice)
     m.close();
     EXPECT_FALSE(m);
     EXPECT_FALSE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }
 
 TEST(magic_open_close_test, open_magic_using_flags_container)
@@ -65,6 +71,7 @@ TEST(magic_open_close_test, open_magic_using_flags_container)
         magic::flags::mime,
         magic::flags::continue_search
     });
-    EXPECT_TRUE(m);
+    EXPECT_FALSE(m);
     EXPECT_TRUE(m.is_open());
+    EXPECT_FALSE(m.is_valid());
 }